Break test file into smaller modules
