+++ package.json
@@ -48,7 +48,7 @@
   },
   "scripts": {
     "test": "npm run mocha && npm run lint",
-    "mocha": "mocha --reporter spec --slow 400 --timeout 8000",
+    "mocha": "mocha test/index.js --reporter spec --slow 400 --timeout 8000",
     "lint": "eslint lib/index.js test/test.js bin/suitcss",
     "watch": "npm run mocha -- --watch"
   },
